move department service of onboarding hr to an query based automatically generated rest endpoint
