+++ onboarding-example/hr/readme.md
@@ -63,7 +63,7 @@ curl -X POST -H 'Accept: application/json' -H 'Content-Type: application/json' -
 Assigns department and manager for the given employee:
 
 ```sh
-curl -X POST -H 'Accept: application/json' -H 'Content-Type: application/json' -d '{"employee" : {"firstName" : "Mark", "lastName" : "Test", "personalId" : "xxx-yy-zzz", "birthDate" : "2012-12-10T14:50:12.123+02:00", "address" : {"country" : "US", "city" : "Boston", "street" : "any street 3", "zipCode" : "10001"}}}' http://localhost:8081/department
+curl -X POST -H 'Accept: application/json' -H 'Content-Type: application/json' -d '{"employees" : [{"firstName" : "Mark", "lastName" : "Test", "personalId" : "xxx-yy-zzz", "birthDate" : "2012-12-10T14:50:12.123+02:00", "address" : {"country" : "US", "city" : "Boston", "street" : "any street 3", "zipCode" : "10001"}}] }' http://localhost:8081/Department
 ```
 
 ## Deployment to OpenShift
