+++ onboarding-example/hr/src/main/java/org/kie/kogito/examples/hr/Employee.java
@@ -67,4 +67,24 @@ public class Employee implements java.io.Serializable {
 		this.birthDate = birthDate;
 		this.address = address;
 	}
+
+	private java.lang.String manager;
+
+	public java.lang.String getManager() {
+		return manager;
+	}
+
+	public void setManager(java.lang.String manager) {
+		this.manager = manager;
+	}
+
+	private java.lang.String department;
+
+	public java.lang.String getDepartment() {
+		return department;
+	}
+
+	public void setDepartment(java.lang.String department) {
+		this.department = department;
+	}
 }
