+++ onboarding-example/hr/src/main/resources/META-INF/kmodule.xml
@@ -1,1 +1,9 @@
-<kmodule xmlns="http://www.drools.org/xsd/kmodule" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"/>
+<?xml version="1.0" encoding="UTF-8"?>
+<kmodule xmlns="http://jboss.org/kie/6.0.0/kmodule">
+    <kbase name="departmentKB" packages="org.kie.kogito.examples.hr.department">
+        <ksession name="departmentKS"/>
+    </kbase>
+    <kbase name="employeesKB" packages="org.kie.kogito.examples.hr.employees">
+        <ksession name="employeesKS" default="true"/>
+    </kbase>
+</kmodule>
