--- onboarding-example/hr/src/main/resources/org/kie/kogito/examples/hr/department.drl
@@ -1,26 +1,0 @@
-package org.kie.kogito.examples.hr
-
-
-rule "US Employee" when
-    $d: DepartmentModel(employee.address.country == "US")
-then
-    modify($d) { setDepartment( "US12345" ) };
-end
-
-rule "UK Employee" when
-    $d: DepartmentModel(employee.address.country == "UK")
-then
-    modify($d) { setDepartment( "UK9876" ) };
-end
-
-rule "US Department" when
-    $d: DepartmentModel(department == "US12345")
-then
-    $d.setManager( "John Doe" );
-end
-
-rule "UK Department" when
-    $d: DepartmentModel(department == "UK9876")
-then
-    $d.setManager( "Teresa April" );
-end
