+++ onboarding-example/hr/src/main/resources/org/kie/kogito/examples/hr/employees/employeeValidation.drl
@@ -1,5 +1,7 @@
-package org.kie.kogito.examples.hr
+package org.kie.kogito.examples.hr.employees
 
+import org.kie.kogito.examples.hr.Employee
+import org.kie.kogito.examples.hr.EmployeeValidationModel
 
 rule "Employee exists"
